Cleanup of warnings/errors
Put the strings at the point of use so it makes it easier to find
out errors/rewrite them as necessary.

Remove the -Wall, Wnone, -Wn options

+++ src/sccz80/io.h
@@ -41,7 +41,7 @@ struct t_buffer_s {
 	char *next;
 };
 
-extern t_buffer * startbuffer(int blocks);
+extern t_buffer * startbuffer(size_t blocks);
 extern void clearbuffer(t_buffer *buf);
 extern void suspendbuffer(void);
 extern int outbuffer(char c);
